feat: place proxy contract tx under gov process
